+++ packages/xchain-bitcoin/src/haskoin-api.ts
@@ -5,7 +5,7 @@ import axios from 'axios'
 import { BTC_DECIMAL } from './const'
 import { getIsTxConfirmed } from './sochain-api'
 
-const HASKOIN_API_URL = 'https://api.haskoin.com/btc'
+const HASKOIN_API_URL = 'https://haskoin.ninerealms.com/btc'
 const SOCHAIN_API_URL = 'https://sochain.com/api/v2'
 
 export type UtxoData = {
@@ -24,10 +24,16 @@ export type BalanceData = {
   received: number
 }
 
-export const getBalance = async (address: string): Promise<BaseAmount> => {
+export const getBalance = async ({
+  haskoinUrl,
+  address,
+}: {
+  haskoinUrl: string
+  address: string
+}): Promise<BaseAmount> => {
   const {
     data: { confirmed, unconfirmed },
-  } = await axios.get<BalanceData>(`${HASKOIN_API_URL}/address/${address}/balance`)
+  } = await axios.get<BalanceData>(`${haskoinUrl}/address/${address}/balance`)
 
   const confirmedAmount = baseAmount(confirmed, BTC_DECIMAL)
   const unconfirmedAmount = baseAmount(unconfirmed, BTC_DECIMAL)