+++ packages/xchain-cosmos/src/cosmos/sdk-client.ts
@@ -16,6 +16,7 @@ import {
   RPCResponse,
   RPCTxSearchResult,
   SearchTxParams,
+  TransferOfflineParams,
   TransferParams,
   TxHistoryResponse,
   TxResponse,
@@ -186,8 +187,7 @@ export class CosmosSDKClient {
     return (await axios.get<TxResponse>(`${this.server}/txs/${hash}`)).data
   }
 
-  async transfer({
-    privkey,
+  unsignedStdTxGet({
     from,
     to,
     amount,
@@ -197,7 +197,7 @@ export class CosmosSDKClient {
       amount: [],
       gas: '200000',
     },
-  }: TransferParams): Promise<BroadcastTxCommitResult> {
+  }: TransferParams) {
     this.setPrefix()
 
     const msg: Msg = [
@@ -214,12 +214,27 @@ export class CosmosSDKClient {
     ]
     const signatures: StdTxSignature[] = []
 
-    const unsignedStdTx = StdTx.fromJSON({
+    return StdTx.fromJSON({
       msg,
       fee,
       signatures,
       memo,
     })
+  }
+
+  async transfer({
+    privkey,
+    from,
+    to,
+    amount,
+    asset,
+    memo = '',
+    fee = {
+      amount: [],
+      gas: '200000',
+    },
+  }: TransferParams): Promise<BroadcastTxCommitResult> {
+    const unsignedStdTx = this.unsignedStdTxGet({ privkey, from, to, amount, asset, memo, fee })
 
     return this.signAndBroadcast(unsignedStdTx, privkey, AccAddress.fromBech32(from))
   }
@@ -241,5 +256,24 @@ export class CosmosSDKClient {
     )
 
     return (await auth.txsPost(this.sdk, signedStdTx, 'sync')).data
+  }
+
+  async transferSignedOffline({
+    privkey,
+    from,
+    from_account_number = '0',
+    from_sequence = '0',
+    to,
+    amount,
+    asset,
+    memo = '',
+    fee = {
+      amount: [],
+      gas: '200000',
+    },
+  }: TransferOfflineParams): Promise<StdTx> {
+    const unsignedStdTx = this.unsignedStdTxGet({ privkey, from, to, amount, asset, memo, fee })
+
+    return auth.signStdTx(this.sdk, privkey, unsignedStdTx, from_account_number, from_sequence)
   }
 }