+++ packages/xchain-doge/src/sochain-api.ts
@@ -14,7 +14,7 @@ import {
 } from './types/sochain-api-types'
 import { DOGE_DECIMAL } from './utils'
 
-const DEFAULT_SUGGESTED_TRANSACTION_FEE = 100000
+const DEFAULT_SUGGESTED_TRANSACTION_FEE = 500000
 
 const toSochainNetwork = (network: Network): string => {
   switch (network) {
@@ -25,8 +25,26 @@ const toSochainNetwork = (network: Network): string => {
   }
 }
 
-export const getSendTxUrl = ({ sochainUrl, network }: AddressParams) => {
-  return `${sochainUrl}/send_tx/${toSochainNetwork(network)}`
+export const getSendTxUrl = ({ sochainUrl, network, token = null }: AddressParams) => {
+  // if (network === 'mainnet') {
+  //   if (token) {
+  //     return `https://api.blockcypher.com/v1/doge/main/txs/push?token=${token}`
+  //   }
+
+  //   return `https://api.blockcypher.com/v1/doge/main/txs/push`
+  // } else {
+  //   return `${sochainUrl}/send_tx/${toSochainNetwork(network)}`
+  // }
+
+  if (network === 'mainnet') {
+    if (token) {
+      return `https://api.blockchair.com/dogecoin/push/transaction?key=${token}`
+    }
+
+    return `https://api.blockchair.com/dogecoin/push/transaction`
+  } else {
+    return `${sochainUrl}/send_tx/${toSochainNetwork(network)}`
+  }
 }
 
 /**
@@ -77,7 +95,9 @@ export const getBalance = async ({ sochainUrl, network, address }: AddressParams
   const url = `${sochainUrl}/get_address_balance/${toSochainNetwork(network)}/${address}`
   const response = await axios.get(url)
   const balanceResponse: SochainResponse<DogeGetBalanceDTO> = response.data
-  const netAmt = balanceResponse.data.balance
+  const confirmed = assetAmount(balanceResponse.data.confirmed_balance, DOGE_DECIMAL)
+  const unconfirmed = assetAmount(balanceResponse.data.unconfirmed_balance, DOGE_DECIMAL)
+  const netAmt = confirmed.amount().plus(unconfirmed.amount())
   const result = assetToBase(assetAmount(netAmt, DOGE_DECIMAL))
   return result
 }
@@ -131,7 +151,7 @@ export const getSuggestedTxFee = async (): Promise<number> => {
   //Note: sochain does not provide fee rate related data
   try {
     const response = await axios.get('https://api.blockcypher.com/v1/doge/main')
-    return response.data.low_fee_per_kb / 1000 // feePerKb to feePerByte
+    return response.data.high_fee_per_kb / 1000 // feePerKb to feePerByte
   } catch (error) {
     return DEFAULT_SUGGESTED_TRANSACTION_FEE
   }