+++ packages/xchain-litecoin/CHANGELOG.md
@@ -1,3 +1,13 @@
+# v.0.7.1 (2022-01-10)
+
+- upgrade axios dependency to latest
+
+# v.0.7.0 (2021-12-29)
+
+## Breaking change
+
+- Add stagenet environment handling for `Network` and `BaseXChainClient` changes client to default to mainnet values until stagenet is configured.
+
 # v.0.6.11 (2021-12-13)
 
 - export `scanUTXO` function